backend/requirements.txt
boxsdk==1.5.5
celery==4.2.1
django-cors-headers==2.4.0
django==2.1.2
djangorestframework-jwt==1.11.0 
djangorestframework==3.8.2
dropbox==9.1.0
google-api-python-client==1.7.4
google-auth-httplib2==0.0.3
google-auth==1.5.1
mysqlclient==1.3.13
oauth2client==4.1.3
onedrivesdk==1.1.8
python-dateutil==2.7.3
redis==2.10.6
requests==2.20.0
Werkzeug==0.14.1
PyDrive==1.3.1
